chore(battery): move out adc logic from app.c
