+++ apps/battery-monitor/meson.build
@@ -1,5 +1,6 @@
 battery_monitor_inc = include_directories('include')
 battery_monitor_src = files(
+  'src' / 'adc.c',
   'src' / 'app.c',
   'src' / 'interrupts.c',
   'src' / 'led.c',
